Allow to use Contribution conditions for Membership End Date (Cron) trigger
